Lab_5/requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
typesense
langchain_openai
langgraph
beautifulsoup4 
black 
tiktoken 
openai 
unstructured 
nltk 
fastapi 
jinja2 
uvicorn 
streamlit 
streamlit-chat 
tqdm 
isort
langchainhub
firecrawl-py
langchain-community
langsmith
pillow
langchain-pinecone
certifi
langchain-chroma
python-dotenv
langchain-tavily 